+++ tests/unit/style/stylesheets.rs
@@ -249,8 +249,8 @@ fn test_parse_stylesheet() {
 
 struct CSSError {
     pub url : ServoUrl,
-    pub line: usize,
-    pub column: usize,
+    pub line: u32,
+    pub column: u32,
     pub message: String
 }
 
@@ -275,7 +275,7 @@ impl ParseErrorReporter for CSSInvalidErrorReporterTest {
                         line_number_offset: u64) {
 
         let location = input.source_location(position);
-        let line_offset = location.line + line_number_offset as usize;
+        let line_offset = location.line + line_number_offset as u32;
 
         let mut errors = self.errors.lock().unwrap();
         errors.push(
